Fix subreddit multireddit top bars.
It previously would take `/top`, and then add *only* the subreddit to it,
rather than the subreddit and multipath.  Now, if the multipath is specified
but not the subreddit (IE, `/m/multi`), the subreddit is added but not the
multi (IE, `/r/subreddit/m/multi`).  However, if the multipath isn't specified,
everything is added: `/top` becomes `/r/subreddit/m/multi/top`, not
`/r/subreddit/top`.